fix(twitter): fix rerendering TwitterList every a player typed text (#518)
