+++ phone/src/apps/twitter/components/TwitterContainer.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react';
-import { TwitterApp } from './TwitterApp';
+import TwitterApp from './TwitterApp';
 import { LoadingSpinner } from '@ui/components/LoadingSpinner';
 
 const TwitterContainer = () => {
